Move main() out into separate file
As a consequence, move the IwyuAction declaration out into iwyu.h so we
can share it between iwyu_main.cc and iwyu.cc.

This should make it easier to link parts of IWYU into other targets.
